% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationPulmonDB.R
\name{annotationPulmonDB}
\alias{annotationPulmonDB}
\title{This gives you manually curated annotation for the samples}
\usage{
annotationPulmonDB(id, output = 1)
}
\arguments{
\item{id}{A value or character vector with GSEs (GEO id).}

\item{output}{todo}
}
\value{
This is the result.
}
\description{
Description.
}
\details{
Details.
}
\examples{
## Example 1
args(annotationPulmonDB)

}
\seealso{
[genesPulmonDB()]
}
\concept{leo}
