% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesPulmonDB.R
\name{genesPulmonDB}
\alias{genesPulmonDB}
\title{Homogenized values in PulmonDB}
\usage{
genesPulmonDB(gene, id)
}
\arguments{
\item{gene}{A value or character vector with gene names. It also accepts
'all' for downloading every available gene.}

\item{id}{A value or character vector with GSEs (GEO id).}
}
\description{
This gives you a S3 object with homogenized values and annotation
per contrast.
}
\details{
Homogenized values are contrasted with a control reference, we have
performed normalization using RMA for Affymetrix platform and
loess normalization for non-Affymetrix.
Rownames are genes and colnames are contrast (GSMXX1.ch1-vs-GSMXX2.ch1),
the annotation is
GSM18413.ch1-vs-GSM18403.ch1       EMPHYSEMA_vs_HEALTHY/CONTROL
}
\examples{

a <- genesPulmonDB(c("MMP1","JUND"), c("GSE1122"))
#annotation
colData(a)
#values
assays(a)$values
assay(a)
all_genes <- genesPulmonDB("all","")

\dontrun{
## a comment
genesPulmonDB("MMP1","JUND", c("GSE1122"))
}

}
\concept{pulmondb}
