% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationPulmonDB.R
\name{annotationPulmonDB}
\alias{annotationPulmonDB}
\title{Annotation from PulmonDB}
\source{
\url{http://pulmondb.liigh.unam.mx/}
}
\usage{
annotationPulmonDB(id, output = "contrast")
}
\arguments{
\item{id}{A value or character vector with GSEs (GEO id).}

\item{output}{"contrast" or "sample". Default is "contrast" but it also can
be changed to return annotation per "sample"}
}
\value{
This is the result.
}
\description{
This gives you manually curated annotation for GSE available in PulmonDB.
}
\details{
Details.
}
\examples{
## Example 1
args(annotationPulmonDB)

}
\seealso{
[genesPulmonDB()]
}
\concept{leo}
