% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesPulmonDB.R
\name{genesPulmonDB}
\alias{genesPulmonDB}
\title{Homogenized values in PulmonDB}
\usage{
genesPulmonDB(gene, id)
}
\arguments{
\item{gene}{A character vector with gene ids.}

\item{id}{A GEO id.}
}
\description{
This gives you a S3 object with homogenized values and annotation
per contrast.
}
\details{
Homogenized values are contrasted with a control reference,
annotation will be have
}
\examples{

a <- genesPulmonDB(c("MMP1","JUND"), c("GSE1122"))
#annotation
colData(a)
#values
assays(a)$values
assay(a)

\dontrun{
## a comment
genesPulmonDB(c("MMP1","JUND"), c("GSE1122"))
}

}
\seealso{
Other leo: \code{\link{annotationPulmonDB}}
}
\concept{leo}
